// src/types/api.generated.ts
export interface components {
    schemas: {
        User: {
            id: string;
            name: string;
            email: string;
            avatar?: string;
            /** @enum {string} */
            role?: "admin" | "user";
        };
        LoginRequest: {
            /** Format: email */
            email: string;
            password: string;
        };
        AuthResponse: {
            user?: components["schemas"]["User"];
            token?: string;
        };
        UserListResponse: {
            data?: components["schemas"]["User"][];
            total?: number;
        };
        Task: {
            id: string;
            title: string;
            description?: string;
            /** @enum {string} */
            status: "todo" | "doing" | "done";
            /**
             * @default medium
             * @enum {string}
             */
            priority: "low" | "medium" | "high";
            /** Format: date-time */
            createdAt?: string;
        };
        TaskCreateDto: {
            title: string;
            description?: string;
            /**
             * @default todo
             * @enum {string}
             */
            status: "todo" | "doing" | "done";
            /**
             * @default medium
             * @enum {string}
             */
            priority: "low" | "medium" | "high";
        };
        TaskUpdateDto: {
            title?: string;
            description?: string;
            /** @enum {string} */
            status?: "todo" | "doing" | "done";
            /** @enum {string} */
            priority?: "low" | "medium" | "high";
        };
        TaskListResponse: {
            data?: components["schemas"]["Task"][];
            total?: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}